Split out `iterator` as a separate module.
