+++ pg_search/src/index/reader/iterator.rs
@@ -1,0 +1,151 @@
+// Copyright (c) 2023-2025 ParadeDB, Inc.
+//
+// This file is part of ParadeDB - Postgres for Search and Analytics
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with this program. If not, see <http://www.gnu.org/licenses/>.
+
+use crate::index::reader::index::enable_scoring;
+use std::sync::OnceLock;
+use tantivy::query::{Query, Scorer};
+use tantivy::{DocAddress, DocId, DocSet, Score, Searcher, SegmentOrdinal, SegmentReader};
+
+pub struct DeferredScorer {
+    query: Box<dyn Query>,
+    need_scores: bool,
+    segment_reader: SegmentReader,
+    searcher: Searcher,
+    scorer: OnceLock<Box<dyn Scorer>>,
+}
+
+impl DeferredScorer {
+    pub fn new(
+        query: Box<dyn Query>,
+        need_scores: bool,
+        segment_reader: SegmentReader,
+        searcher: Searcher,
+    ) -> Self {
+        Self {
+            query,
+            need_scores,
+            segment_reader,
+            searcher,
+            scorer: Default::default(),
+        }
+    }
+
+    #[track_caller]
+    #[inline(always)]
+    fn scorer_mut(&mut self) -> &mut Box<dyn Scorer> {
+        self.scorer();
+        self.scorer
+            .get_mut()
+            .expect("deferred scorer should have been initialized")
+    }
+
+    #[track_caller]
+    #[inline(always)]
+    fn scorer(&self) -> &dyn Scorer {
+        self.scorer.get_or_init(|| {
+            let weight = self
+                .query
+                .weight(enable_scoring(self.need_scores, &self.searcher))
+                .expect("weight should be constructable");
+
+            weight
+                .scorer(&self.segment_reader, 1.0)
+                .expect("scorer should be constructable")
+        })
+    }
+}
+
+impl DocSet for DeferredScorer {
+    #[inline(always)]
+    fn advance(&mut self) -> DocId {
+        self.scorer_mut().advance()
+    }
+
+    #[inline(always)]
+    fn doc(&self) -> DocId {
+        self.scorer().doc()
+    }
+
+    fn size_hint(&self) -> u32 {
+        self.scorer().size_hint()
+    }
+}
+
+impl Scorer for DeferredScorer {
+    #[inline(always)]
+    fn score(&mut self) -> Score {
+        self.scorer_mut().score()
+    }
+}
+
+pub struct ScorerIter {
+    deferred: DeferredScorer,
+    segment_ord: SegmentOrdinal,
+    segment_reader: SegmentReader,
+}
+
+impl ScorerIter {
+    pub fn new(
+        scorer: DeferredScorer,
+        segment_ord: SegmentOrdinal,
+        segment_reader: SegmentReader,
+    ) -> Self {
+        Self {
+            deferred: scorer,
+            segment_ord,
+            segment_reader,
+        }
+    }
+}
+
+impl Iterator for ScorerIter {
+    type Item = (Score, DocAddress);
+
+    fn next(&mut self) -> Option<Self::Item> {
+        loop {
+            let doc_id = self.deferred.doc();
+
+            if doc_id == tantivy::TERMINATED {
+                // we've read all the docs
+                return None;
+            } else if self
+                .segment_reader
+                .alive_bitset()
+                .map(|alive_bitset| alive_bitset.is_alive(doc_id))
+                // if there's no alive_bitset, the doc is alive
+                .unwrap_or(true)
+            {
+                // this doc is alive
+                let score = self.deferred.score();
+                let this = (score, DocAddress::new(self.segment_ord, doc_id));
+
+                // move to the next doc for the next iteration
+                self.deferred.advance();
+
+                // return the live doc
+                return Some(this);
+            }
+
+            // this doc isn't alive, move to the next doc and loop around
+            self.deferred.advance();
+        }
+    }
+
+    fn size_hint(&self) -> (usize, Option<usize>) {
+        (0, Some(self.deferred.size_hint() as usize))
+    }
+}
