+++ pg_search/src/index/reader/mod.rs
@@ -1,2 +1,3 @@
 pub mod index;
+pub mod iterator;
 pub mod segment_component;
